Fix type in tests

+++ tests/create_curve_data_test.py
@@ -46,9 +46,13 @@ def test_create_curve_data(tmp_path):
         data_objects = curve.add_data(
             {
                 "vertexValues": {
-                    "values": np.random.randint(0, curve.n_vertices, curve.n_vertices)
+                    "values": np.random.randint(
+                        0, curve.n_vertices, curve.n_vertices
+                    ).astype(np.uint32)
                 },
-                "cellValues": {"values": np.random.randn(curve.n_cells)},
+                "cellValues": {
+                    "values": np.random.randn(curve.n_cells).astype(np.float64)
+                },
             }
         )
